fix mouse lag
When using a high polling rate mouse on Linux with X11, the mouse cursor
lags. This is due to the fact that Allegro will only process up-to 5
X11 events at a time, and so mouse motion events will pile up and the
mouse cursor will lag. It's possible to fix it without patching Allegro
by setting a custom Allegro X11 input handler that will make sure all
currently queued X11 events are processed.